General function for ingesting GWAS summary statistics into GWAS-VCF format

Create a standardized VCF file

Usage

format_sumstat2vcf(
  path,
  SNP,
  CHR,
  BP,
  A1,
  A2,
  FRQ,
  BETA,
  SE,
  P,
  NCAS,
  N,
  ref_genome,
  id,
  studytype,
  save_path,
  hg19_1kg_phase3,
  trait,
  group_name = NA,
  year = NA,
  author = NA,
  consortium = NA,
  sex = NA,
  pmid = NA,
  population = NA,
  unit = NA,
  subcategory = NA,
  ontology = NA,
  note = NA,
  doi = NA,
  sd = NA,
  nThread,
  ...
)

Arguments

path

the file name with a path which the summary statistics is to be read from.This can be a compressed file as *.tsv.gz

SNP

column name for SNP (rs id). If NULL, rs id inferred from reference genome by MungeSumstat

CHR

column name for chromosome. Required.

BP

column name for base pair positions. Required.

A1

column name for reference allele. If NULL, inferred by MungeSumstat.

A2

column name for alternative allele. Required. A note on the allele flipping check: MungeSumstats infers the effect allele will always be the A2 allele, this is the approach done for IEU GWAS VCF and has such also been adopted here.

FRQ

column name for alternate allele frequency of the SNP.If there is no column for FRQ, set as NULL, inferred from 1kg reference

BETA

column name for effect size estimate relative to the alternative allele. If NULL, MungeSumstats sets impute_beta = TRUE

SE

column name for standard error of effect size estimate. If NULL, then MungeSumstats sets impute_se = TRUE

P

column name for unadjusted p-value of effect estimate. Required.

NCAS

column name for number of cases, if studytype is "Binary",Required.If there is no column for NCAS, input an integer number. if studytype is "Continuous", NA.

N

column name for total sample size. Required. If there is no column for sample size, then input an integer, MungeSumstats sets compute_n to this integer value the N (sample size) for every SNP in the dataset

ref_genome

name of the reference genome used for the GWAS ("GRCh37" or "GRCh38") Argument is case-insensitive. Required.

id

name of the vcf output file and used in the gwas catalogue as ID identifier

studytype

"Continuous" for quantitative trait and "Binary" for binary trait.

save_path

the path which the vcf will be written out. End with "/"

hg19_1kg_phase3

if FRQ is NULL, annotate the FRQ with the 1000 genomes phase 3 (hg19/GRCh37) allele frequencies and index by "/cloud-home/U1006974/GiTIPs/data/reference/ALL.wgs.phase3_shapeit2_mvncall_integrated_v5c.20130502.sites.AF.tsv"

trait

trait name. Required.

group_name

group name. Optional. "public" or "private" or ...

year

Optional.

author

Optional.

consortium

Optional.

sex

Optional.

pmid

Optional.

population

Optional.

unit

Optional.

subcategory

Optional.

ontology

Optional.

note

Optional.

doi

Optional.

sd

Optional.

nThread

Number of threads to use for parallel processes

...

parameters from MungeSumstats. See the reference below

Value

one .vcf.gz file, one .vcf.gz.tbi file, one a log file .log, and one catalog file, .clg. Note: The majority of VCF handling bioinformatics libraries use a .vcf.gz suffix, even for block gziped output. However, VariantAnnotation::writeVcf() even with index = TRUE does not support this and forceably sets the suffix to .bgz. See https://github.com/Bioconductor/VariantAnnotation/issues/35. So the current solution is to rename from vcf.bgz to vcf.gz

References

https://github.com/MRCIEU/gwas-vcf-specification

https://neurogenomics.github.io/MungeSumstats/reference/index.html