Function for obtaining top hits from a GWAS vcf file with and without clumping

Obtain top hits from a GWAS vcf file with/without clumping Note: gwasglue::clump_gwasvcf function clumping using 1kg reference genome or other reference genome UKbiobank, the issue is that if a top variant is missing from the reference genome, it will it will "Removing top variants absence from LD reference panel", which is not ideal. As GWAS summary statistic are from imputed genome, for sure it will cover more SNPs than the unimputed reference genome. The current solution is that get the top SNP in each chromosome (unclumped_top), then do the clumping using reference genome (clumped_df), if (clumped_df) miss a chromosome in (unclumped_top), combine it into (clumped_df)

Usage

tophits_vcf(
  path,
  clump_p = 5e-08,
  clump_r2 = 0.001,
  clump_kb = 10000,
  ldref,
  plink_bin = genetics.binaRies::get_plink_binary(),
  nThread
)

Arguments

path

Specify the full path starting with /cloud-home/ to the vcf file, including the file name.

clump_p

p value threshold to filter. Default = 5e-8

clump_r2

Clumping r2 threshold. Default: 0.001

clump_kb

Clumping kb window. Default: 10000. (very strict)

ldref

LD reference data. Normally Europeans from the 1000 genomes reference panel. Default:"/cloud-home/U1006974/NeuroID/data/1kg/EUR"

plink_bin

specify path to plink binary. Default = "genetics.binaRies::get_plink_binary()"

nThread

Number of threads to use for parallel processes

Value

a data.frame for top hits from a GWAS vcf file with/without clumping

References

https://mrcieu.github.io/gwasglue/reference/clump_gwasvcf.html